Cleaned up the application.scala into subpackages
