+++ jvm/rtbtc/app/utils/Events.scala
@@ -1,0 +1,6 @@
+package utils
+
+object Events {
+  case class GraphEventRegistration(topic: String)
+  case class GraphEventPoint(topic: String, sell: Double, buy: Double, volume: Double)
+}
